Include Kotlin DSL samples in build-dashboard plugin doc
This change also improves the sample test to ensure the buildDashboard
task is actually listed.
