+++ subprojects/docs/src/samples/buildDashboard/buildDashboard.out
@@ -1,0 +1,3 @@
+Reporting tasks
+---------------
+buildDashboard - Generates a dashboard of all the reports produced by this build.
