+++ subprojects/docs/src/samples/buildDashboard/buildDashboard.sample.conf
@@ -1,0 +1,13 @@
+commands: [{
+    execution-subdirectory: groovy
+    executable: gradle
+    args: tasks
+    expected-output-file: buildDashboard.out
+    allow-additional-output: true
+},{
+    execution-subdirectory: kotlin
+    executable: gradle
+    args: tasks
+    expected-output-file: buildDashboard.out
+    allow-additional-output: true
+}]
