+++ subprojects/docs/src/samples/buildDashboard/groovy/config/codenarc/codenarc.xml
