+++ subprojects/docs/src/samples/buildDashboard/groovy/settings.gradle
