+++ subprojects/docs/src/samples/buildDashboard/groovy/src/main/groovy/org/gradle/sample/GroovyPerson.groovy
