+++ subprojects/docs/src/samples/buildDashboard/groovy/src/main/java/org/gradle/sample/Person.java
@@ -1,0 +1,15 @@
+package org.gradle.sample;
+
+import java.lang.String;
+
+class Person {
+    private String name;
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+}
