+++ subprojects/docs/src/samples/buildDashboard/kotlin/settings.gradle.kts
@@ -1,0 +1,1 @@
+rootProject.name = "build-dashboard"
