+++ subprojects/docs/src/samples/buildDashboard/kotlin/src/main/groovy/org/gradle/sample/GroovyPerson.groovy
@@ -1,0 +1,5 @@
+package org.gradle.sample
+
+class GroovyPerson {
+    def String name
+}
