--- subprojects/docs/src/samples/buildDashboard/useBuildDashboardPlugin.sample.conf
@@ -1,2 +1,0 @@
-executable: gradle
-args: tasks
